Use DataSetSampleActions everywhere samples are added to the dataset
